requirements.txt
Pillow==10.2.0
Requests==2.31.0
streamlit==1.31.1
streamlit_lottie==0.0.5
